+++ jgiven-android/src/main/java/com.tngtech.jgiven.integration.android/AndroidStageClassCreator.java
@@ -3,12 +3,12 @@ package com.tngtech.jgiven.integration.android;
 import android.support.test.InstrumentationRegistry;
 import android.support.v4.content.ContextCompat;
 
-import com.tngtech.jgiven.impl.ByteBuddyStageCreator;
+import com.tngtech.jgiven.impl.ByteBuddyStageClassCreator;
 
 import net.bytebuddy.android.AndroidClassLoadingStrategy;
 import net.bytebuddy.dynamic.loading.ClassLoadingStrategy;
 
-public class AndroidStageCreator extends ByteBuddyStageCreator {
+public class AndroidStageClassCreator extends ByteBuddyStageClassCreator {
 
     @Override
     protected ClassLoader getClassLoader(Class<?> stageClass) {