+++ jgiven-core/src/main/java/com/tngtech/jgiven/impl/intercept/ByteBuddyMethodInterceptor.java
@@ -3,15 +3,20 @@ package com.tngtech.jgiven.impl.intercept;
 import java.lang.reflect.Method;
 import java.util.concurrent.Callable;
 
+import com.tngtech.jgiven.impl.ByteBuddyStageClassCreator;
+import com.tngtech.jgiven.impl.ByteBuddyStageClassCreator.StepInterceptorGetterSetter;
 import com.tngtech.jgiven.impl.intercept.StepInterceptor.Invoker;
 
 import net.bytebuddy.implementation.bind.annotation.AllArguments;
 import net.bytebuddy.implementation.bind.annotation.BindingPriority;
 import net.bytebuddy.implementation.bind.annotation.DefaultCall;
+import net.bytebuddy.implementation.bind.annotation.FieldProxy;
 import net.bytebuddy.implementation.bind.annotation.Origin;
 import net.bytebuddy.implementation.bind.annotation.RuntimeType;
 import net.bytebuddy.implementation.bind.annotation.SuperCall;
 import net.bytebuddy.implementation.bind.annotation.This;
+
+import static com.tngtech.jgiven.impl.ByteBuddyStageClassCreator.INTERCEPTOR_FIELD_NAME;
 
 /**
  * StepInterceptorImpl that uses ByteBuddy Method interceptor with annotations for intercepting JGiven methods
@@ -19,33 +24,16 @@ import net.bytebuddy.implementation.bind.annotation.This;
  */
 public class ByteBuddyMethodInterceptor {
 
-    /**
-     * The interceptor to delegate method calls.
-     * Can be {@code null} in some cases until the scenario has started
-     */
-    private StepInterceptor interceptor;
-
-    public ByteBuddyMethodInterceptor() {}
-
-    public ByteBuddyMethodInterceptor( StepInterceptor interceptor ) {
-        this.interceptor = interceptor;
-    }
-
-    public void setInterceptor( StepInterceptor interceptor ) {
-        this.interceptor = interceptor;
-    }
-
     @RuntimeType
     @BindingPriority( BindingPriority.DEFAULT * 3 )
     public Object interceptSuper( @SuperCall final Callable<?> zuper, @This final Object receiver, @Origin Method method,
-            @AllArguments final Object[] parameters )
+            @AllArguments final Object[] parameters,
+            @FieldProxy( INTERCEPTOR_FIELD_NAME ) StepInterceptorGetterSetter stepInterceptorGetter )
             throws Throwable {
 
-        if( handleSetStepInterceptor( method, parameters ) ) {
-            return null;
-        }
+        StepInterceptor interceptor = (StepInterceptor) stepInterceptorGetter.getValue();
 
-        if (interceptor == null) {
+        if( interceptor == null ) {
             return zuper.call();
         }
 
@@ -62,14 +50,13 @@ public class ByteBuddyMethodInterceptor {
     @RuntimeType
     @BindingPriority( BindingPriority.DEFAULT * 2 )
     public Object interceptDefault( @DefaultCall final Callable<?> zuper, @This final Object receiver, @Origin Method method,
-            @AllArguments final Object[] parameters )
+            @AllArguments final Object[] parameters,
+            @FieldProxy( INTERCEPTOR_FIELD_NAME ) StepInterceptorGetterSetter stepInterceptorGetter )
             throws Throwable {
 
-        if( handleSetStepInterceptor( method, parameters ) ) {
-            return null;
-        }
+        StepInterceptor interceptor = (StepInterceptor) stepInterceptorGetter.getValue();
 
-        if (interceptor == null) {
+        if( interceptor == null ) {
             return zuper.call();
         }
 
@@ -84,15 +71,14 @@ public class ByteBuddyMethodInterceptor {
     }
 
     @RuntimeType
-    public Object intercept( @This final Object receiver, @Origin final Method method, @AllArguments final Object[] parameters )
+    public Object intercept( @This final Object receiver, @Origin final Method method,
+                             @AllArguments final Object[] parameters,
+                             @FieldProxy( INTERCEPTOR_FIELD_NAME ) StepInterceptorGetterSetter stepInterceptorGetter)
             throws Throwable {
         // this intercepted method does not have a non-abstract super method
+        StepInterceptor interceptor = (StepInterceptor) stepInterceptorGetter.getValue();
 
-        if( handleSetStepInterceptor( method, parameters ) ) {
-            return null;
-        }
-
-        if (interceptor == null) {
+        if( interceptor == null ) {
             return null;
         }
 
@@ -105,14 +91,6 @@ public class ByteBuddyMethodInterceptor {
 
         };
         return interceptor.intercept( receiver, method, parameters, invoker );
-    }
-
-    private boolean handleSetStepInterceptor( Method method, final Object[] parameters ) {
-        if( method.getName().equals( "setStepInterceptor" ) && method.getDeclaringClass().equals( StageInterceptorInternal.class ) ) {
-            setInterceptor( (StepInterceptor) parameters[0] );
-            return true;
-        }
-        return false;
     }
 
 }