+++ jgiven-core/src/main/java/com/tngtech/jgiven/impl/intercept/StageInterceptorInternal.java
@@ -1,5 +1,5 @@
 package com.tngtech.jgiven.impl.intercept;
 
 public interface StageInterceptorInternal {
-    void setStepInterceptor( StepInterceptor stepInterceptor );
+    void __jgiven_setStepInterceptor( StepInterceptor stepInterceptor );
 }