+++ jgiven-spring/src/main/java/com/tngtech/jgiven/integration/spring/SpringStageCreator.java
@@ -7,8 +7,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.util.ClassUtils;
 
-import com.tngtech.jgiven.impl.ByteBuddyStageCreator;
-import com.tngtech.jgiven.impl.intercept.StageInterceptorInternal;
+import com.tngtech.jgiven.impl.ByteBuddyStageClassCreator;
+import com.tngtech.jgiven.impl.DefaultStageCreator;
+import com.tngtech.jgiven.impl.CachingStageClassCreator;
 import com.tngtech.jgiven.impl.intercept.StepInterceptor;
 
 /**
@@ -27,18 +28,25 @@ import com.tngtech.jgiven.impl.intercept.StepInterceptor;
  * <strong>The SpringStageCreator is stateful, and thus should use "prototype" scope</strong>
  * @since 0.8.0
  */
-public class SpringStageCreator extends ByteBuddyStageCreator {
+public class SpringStageCreator extends DefaultStageCreator {
 
     private static final Logger log = LoggerFactory.getLogger( SpringStageCreator.class );
 
+    private static final CachingStageClassCreator fallBackStageClassCreator = new CachingStageClassCreator(
+        new ByteBuddyStageClassCreator() );
+
     @Autowired
     private ApplicationContext applicationContext;
+
+    public SpringStageCreator() {
+        super( fallBackStageClassCreator );
+    }
 
     @Override
     public <T> T createStage( Class<T> stageClass, StepInterceptor stepInterceptor ) {
         try {
             T bean = applicationContext.getBean( stageClass );
-            ( (StageInterceptorInternal) bean ).setStepInterceptor( stepInterceptor );
+            setStepInterceptor( bean, stepInterceptor);
             return bean;
         } catch( NoSuchBeanDefinitionException nbe ) {
             return super.createStage( stageClass, stepInterceptor );
@@ -48,7 +56,7 @@ public class SpringStageCreator extends ByteBuddyStageCreator {
                 cce );
             return super.createStage( stageClass, stepInterceptor );
         } catch( Exception e ) {
-            log.error( "Error while trying to get the Spring bean for stage class "+stageClass, e );
+            log.error( "Error while trying to get the Spring bean for stage class " + stageClass, e );
             return null;
         }