+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 ## New Features
 
 * Thumbnail preview for image attachments added [#274](https://github.com/TNG/JGiven/pull/299)
+* Reworked AbstractReport interface for easier customizability of new reports and better separation of custom flags [#274](https://github.com/TNG/JGiven/pull/299)
+
+# v0.14.1
+
+## Fixed Issues
+
+* Performance: Caching the classes generated with ByteBuddy introduced with v0.14.0 to decrease the memory consumption and performance of creating stage classes [#294](https://github.com/TNG/JGiven/issues/294)
+* OSGi: Using class loader of the stage class instead of the current thread when creating classes with ByteBuddy [#302](https://github.com/TNG/JGiven/issues/302)
+* HTML Report: Long exception messages of failed Scenarios are now wrapped [#292](https://github.com/TNG/JGiven/issues/292)
 
 # v0.14.0